+++ src/threads.js
@@ -2040,7 +2040,7 @@ Process.prototype.doInsertInList = function (element, index, list) {
         return null;
     }
     if (index instanceof Array) {
-        if (index[0] === 'any') {
+        if (index[0] === 'random') {
             idx = this.reportBasicRandom(1, list.length() + 1);
         } else if (index[0] === 'last') {
             idx = list.length() + 1;