+++ HISTORY.md
@@ -7,6 +7,9 @@
 * **Documentation Updates:**
 * **Translation Updates:**
 
+### 2023-03-10
+* new dev version
+
 ## 8.2.1:
 * **Notable Fixes:**
     * fixed #3190 - unable to search blocks in the stage